refactor: extract DocumentCard edit modal into new component
